pyproject.toml
[tool.poetry]
name = "xgbpid"
version = "0.1.0"
description = ""
authors = ["mresath <mresath@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
xgboost = "^3.2.0"
polars = "^1.38.1"
numpy = "^2.4.2"
scipy = "^1.17.1"
pydantic = "^2.12.5"
pyyaml = "^6.0.3"
streamlit = "^1.45.0"
plotly = "^6.1.0"
supabase = "^2.28.0"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
ruff = "^0.15.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
